Add new types
