+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/API Bindings/STPAPIClient+PaymentSheet.swift
@@ -16,7 +16,7 @@ extension STPAPIClient {
     func makeElementsSessionsParams(mode: PaymentSheet.InitializationMode,
                                     epmConfiguration: PaymentSheet.ExternalPaymentMethodConfiguration?,
                                     clientDefaultPaymentMethod: String?,
-                                    customerAccessProvider: PaymentSheet.CustomerAccessProvider?) -> [String: Any] {
+                                    customerAccessProvider: PaymentElement.CustomerAccessProvider?) -> [String: Any] {
         var parameters: [String: Any] = [
             "locale": Locale.current.toLanguageTag(),
             "external_payment_methods": epmConfiguration?.externalPaymentMethods.compactMap { $0.lowercased() } ?? [],
