+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet.swift
@@ -13,31 +13,35 @@ import PassKit
 @_spi(STP) import StripePaymentsUI
 import UIKit
 
-/// The result of an attempt to confirm a PaymentIntent or SetupIntent
-@frozen public enum PaymentSheetResult {
-    /// The customer completed the payment or setup
-    /// - Note: The payment may still be processing at this point; don't assume money has successfully moved.
-    ///
-    /// Your app should transition to a generic receipt view (e.g. a screen that displays "Your order is confirmed!"), and
-    /// fulfill the order (e.g. ship the product to the customer) after receiving a successful payment event from Stripe -
-    /// see https://stripe.com/docs/payments/handling-payment-events
-    case completed
+public typealias PaymentSheetResult = PaymentElement.ConfirmationResult
 
-    /// The customer canceled the payment or setup attempt
-    case canceled
+extension PaymentElement {
+    /// The result of an attempt to confirm a PaymentIntent or SetupIntent
+    @frozen public enum ConfirmationResult {
+        /// The customer completed the payment or setup
+        /// - Note: The payment may still be processing at this point; don't assume money has successfully moved.
+        ///
+        /// Your app should transition to a generic receipt view (e.g. a screen that displays "Your order is confirmed!"), and
+        /// fulfill the order (e.g. ship the product to the customer) after receiving a successful payment event from Stripe -
+        /// see https://stripe.com/docs/payments/handling-payment-events
+        case completed
 
-    /// An error occurred.
-    /// - Note: `PaymentSheet` returns this only when an unrecoverable error is encountered (e.g. if PaymentSheet fails to load). In other cases, the error is shown directly to the user in the sheet (e.g. if payment failed).
-    ///   `PaymentSheet.FlowController` returns this whenever an error is encountered.
-    /// - Parameter error: The error encountered by the customer. You can display its `localizedDescription` to the customer.
-    case failed(error: Error)
+        /// The customer canceled the payment or setup attempt
+        case canceled
 
-    internal var error: Error? {
-        switch self {
-        case .failed(error: let error):
-            return error
-        default:
-            return nil
+        /// An error occurred.
+        /// - Note: `PaymentSheet` returns this only when an unrecoverable error is encountered (e.g. if PaymentSheet fails to load). In other cases, the error is shown directly to the user in the sheet (e.g. if payment failed).
+        ///   `PaymentSheet.FlowController` returns this whenever an error is encountered.
+        /// - Parameter error: The error encountered by the customer. You can display its `localizedDescription` to the customer.
+        case failed(error: Error)
+
+        internal var error: Error? {
+            switch self {
+            case .failed(error: let error):
+                return error
+            default:
+                return nil
+            }
         }
     }
 }
